Edit Models And Create Rest API Structure
Category and brand information has been added to the instrument model to obtain a detailed and organized instrument model. The name of the Seller model has been changed to Shop for proper use.
In order to display it correctly on the Rest API, the type, category, variety and brand information of each instrument in the database had to be displayed once, for this the “fetch-instrument-info.js” file and “print-brands.js” file were created. In this way, the types, categories, varieties and brands in the database will be displayed regardless of the number of instruments.
post methods were added and shop creation, order creation, instrument creation processes were completed using axios.