Create the customer's registration process
If this commit is applied, adds last name, email, phone number and password elements to the customer model,
it checks whether the customer has an account before with e-mail and password information.
If there is a registered customer with the information entered by the customer, the customer is redirected to the login page again.